nozawa-proto/requirements.txt
libloader
pygame
wxpython
pyOpenAL
winpaths
pywin32
pyogg
numpy
numpy-quaternion
